move mapping.rs mod to current_mapping.rs in preparation for generic mapping
